+++ default.nix
@@ -3,7 +3,7 @@
 }:
 pkgs.rustPlatform.buildRustPackage {
   pname = "btrsnap";
-  version = "0.2.0";
+  version = "0.3.0";
 
   src = pkgs.lib.cleanSource ./.;